requirements.txt
flask
psycopg2-binary
langchain
langchain-community
faiss-cpu
python-dotenv